+++ R/autoplot-cps-spct.r
@@ -65,10 +65,7 @@ cps_plot <- function(spct,
   if (!photobiology::is.cps_spct(spct)) {
     stop("cps_plot() can only plot cps_spct objects.")
   }
-  if (!is.null(geom) && !geom %in% c("area", "line", "spct")) {
-    warning("'geom = ", geom, "' not supported, using default instead.")
-    geom <- NULL
-  }
+  geom <- validate_geom_arg(geom)
   if (is.null(ylim) || !is.numeric(ylim)) {
     ylim <- rep(NA_real_, 2L)
   }
@@ -194,10 +191,18 @@ cps_plot <- function(spct,
     }
   }
 
-  if (!is.null(geom) && geom %in% c("area", "spct")) {
+  if ("col" %in% geom) {
+    plot <- plot + geom_col(fill = "black", colour = NA, alpha = 0.2)
+  }
+  if (any(c("area", "spct") %in% geom)) {
     plot <- plot + geom_spct(fill = "black", colour = NA, alpha = 0.2)
   }
-  plot <- plot + ggplot2::geom_line(na.rm = na.rm)
+  if (!length(geom) || "line" %in% geom) {
+    plot <- plot + ggplot2::geom_line(na.rm = na.rm)
+  }
+  if ("point" %in% geom) {
+    plot <- plot + geom_point()
+  }
   plot <- plot + ggplot2::labs(x = bquote("Wavelength, "*lambda~(nm)),
                                y = s.cps.label)