+++ R/autoplot-response-spct.r
@@ -64,10 +64,7 @@ e_rsp_plot <- function(spct,
     stop("e_Rsp_plot() can only plot response_spct objects.")
   }
   spct[["s.q.response"]] <- NULL
-  if (!is.null(geom) && !geom %in% c("area", "line", "spct")) {
-    warning("'geom = ", geom, "' not supported, using default instead.")
-    geom <- NULL
-  }
+  geom <- validate_geom_arg(geom)
   if (is.null(ylim) || !is.numeric(ylim)) {
     ylim <- rep(NA_real_, 2L)
   }
@@ -214,10 +211,18 @@ e_rsp_plot <- function(spct,
       ggplot2::geom_hline(yintercept = 0, linetype = "dashed", colour = "black")
   }
 
-  if (!is.null(geom) && geom %in% c("area", "spct")) {
+  if ("col" %in% geom) {
+    plot <- plot + geom_col(fill = "black", colour = NA, alpha = 0.2)
+  }
+  if (any(c("area", "spct") %in% geom)) {
     plot <- plot + geom_spct(fill = "black", colour = NA, alpha = 0.2)
   }
-  plot <- plot + ggplot2::geom_line(na.rm = na.rm)
+  if (!length(geom) || "line" %in% geom) {
+    plot <- plot + ggplot2::geom_line(na.rm = na.rm)
+  }
+  if ("point" %in% geom) {
+    plot <- plot + geom_point()
+  }
   plot <- plot + labs(x = bquote("Wavelength, "*lambda~(nm)), y = s.rsp.label)
 
   if (length(annotations) == 1 && annotations == "") {
@@ -339,10 +344,7 @@ q_rsp_plot <- function(spct,
     stop("q_Rsp_plot() can only plot response_spct objects.")
   }
   spct[["s.e.response"]] <- NULL
-  if (!is.null(geom) && !geom %in% c("area", "line", "spct")) {
-    warning("'geom = ", geom, "' not supported, using default instead.")
-    geom <- NULL
-  }
+  geom <- validate_geom_arg(geom)
   if (is.null(ylim) || !is.numeric(ylim)) {
     ylim <- rep(NA_real_, 2L)
   }
@@ -489,10 +491,18 @@ q_rsp_plot <- function(spct,
       ggplot2::geom_hline(yintercept = 0, linetype = "dashed", colour = "black")
   }
 
-  if (!is.null(geom) && geom %in% c("area", "spct")) {
+  if ("col" %in% geom) {
+    plot <- plot + geom_col(fill = "black", colour = NA, alpha = 0.2)
+  }
+  if (any(c("area", "spct") %in% geom)) {
     plot <- plot + geom_spct(fill = "black", colour = NA, alpha = 0.2)
   }
-  plot <- plot + ggplot2::geom_line(na.rm = na.rm)
+  if (!length(geom) || "line" %in% geom) {
+    plot <- plot + ggplot2::geom_line(na.rm = na.rm)
+  }
+  if ("point" %in% geom) {
+    plot <- plot + geom_point()
+  }
   plot <- plot + ggplot2::labs(x = bquote("Wavelength, "*lambda~(nm)),
                                y = s.rsp.label)